+++ SA3/qa_web/tests/test_views.py
@@ -1,0 +1,16 @@
+from django.test import TestCase
+from qa_web.views import index, QuestionDisplayView
+
+
+class ViewTest(TestCase):
+    """This class contains test methods for view functions"""
+
+    def test_index(self):
+        pass
+
+
+class QuestionDisplayViewTest(TestCase):
+    """This class contains test cases for the QuestionDisplayView"""
+
+    def test_get_context(self):
+        pass
